# src/main/java/com/sdchat/ce/sp/complexity/model/BatchItem.java
package com.sdchat.ce.sp.complexity.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class BatchItem {

    private String type;

    private String content;

    private String name;
}