# src/main/java/com/sdchat/ce/sp/complexity/model/BatchResultItem.java
package com.sdchat.ce.sp.complexity.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class BatchResultItem {

    private String name;

    private String type;

    private Boolean success;

    private HintAnalysisSummary analysis;
}