# src/main/java/com/sdchat/ce/sp/complexity/model/BatchValidateRequest.java
package com.sdchat.ce.sp.complexity.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class BatchValidateRequest {

    private List<BatchItem> items;

    @Builder.Default
    private String format = "json";
}