# src/main/java/com/sdchat/ce/sp/complexity/model/BatchValidationResponse.java
package com.sdchat.ce.sp.complexity.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class BatchValidationResponse {

    private Boolean success;

    private List<BatchResultItem> results;

    private HintAnalysisSummary summary;

    private Long processingTimeMs;
}