# src/main/java/com/sdchat/ce/sp/complexity/model/HintParameter.java
package com.sdchat.ce.sp.complexity.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class HintParameter {

    private String name;

    private String description;

    private Boolean required;
}