# src/main/java/com/sdchat/ce/sp/complexity/model/ParseError.java
package com.sdchat.ce.sp.complexity.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class ParseError {

    private String statementId;

    private Integer lineNumber;

    private String errorMessage;

    private String partialText;
}