# src/main/java/com/sdchat/ce/sp/complexity/model/ValidateProcedureFileRequest.java
package com.sdchat.ce.sp.complexity.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;
import org.springframework.web.multipart.MultipartFile;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class ValidateProcedureFileRequest {

    private MultipartFile file;

    private String procedureName;

    @Builder.Default
    private String format = "json";
}