# src/main/java/com/sdchat/ce/sp/complexity/model/ValidateProcedureRequest.java
package com.sdchat.ce.sp.complexity.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class ValidateProcedureRequest {

    private String procedure;

    private String procedureName;

    @Builder.Default
    private String format = "json";
}