# src/main/java/com/sdchat/ce/sp/complexity/model/ValidateSqlRequest.java
package com.sdchat.ce.sp.complexity.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class ValidateSqlRequest {

    private String sql;

    @Builder.Default
    private Boolean includeDetails = true;
}